CDIO/requirements.txt
flask>=3.0.0
selenium>=4.18.0
webdriver-manager>=4.0.1
pymysql>=1.1.0
werkzeug>=3.0.0
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0